react select is moved to a separate component
