+++ src/Components/StylishReactSelect.tsx
@@ -1,17 +1,55 @@
 /* eslint-disable max-len */
-import { DropdownIndicatorProps, GroupBase, components } from 'react-select';
+import classNames from 'classnames';
+import Select, { ClassNamesConfig, StylesConfig } from 'react-select';
+import { typographyStyle } from '../CustomStyles/Typography';
 
-export const DropdownIndicator = (
-  props: JSX.IntrinsicAttributes &
-  DropdownIndicatorProps<unknown, boolean, GroupBase<unknown>>,
-) => {
-  const { menuIsOpen } = props.selectProps;
+type Props = {
+  className: string;
+  options: unknown[];
+  value: unknown;
+  onChange: (e: unknown) => void;
+};
 
-  const caretClass = menuIsOpen ? 'caret-up' : 'caret-down';
+const customClasses: ClassNamesConfig = {
+  menu: () => 'rounded-none',
+  menuList: () => `text-Secondary border py-2 border-Elements bg-white ${typographyStyle.bodyText}`,
+  control: () => 'rounded-none border border-Icons px-3 py-0 shadow-none focus-within:border-Primary hover:border-Secondary',
+  option: state => classNames(
+    'flex h-8 items-center px-3 py-0 hover:bg-Background hover:text-Primary',
+    {
+      'bg-Background text-Primary': state.isSelected,
+    },
+  ),
+  dropdownIndicator: state => classNames('transition-all', {
+    'rotate-180': state.selectProps.menuIsOpen,
+  }),
+};
 
+const customStyles: StylesConfig = {
+  option: state => ({
+    ...state,
+    display: 'flex',
+  }),
+};
+
+export const StylishReactSelect: React.FC<Props> = ({
+  className,
+  options,
+  value,
+  onChange,
+}) => {
   return (
-    <components.DropdownIndicator {...props}>
-      <div className={`${caretClass}`} />
-    </components.DropdownIndicator>
+    <Select
+      options={options}
+      onChange={onChange}
+      value={value}
+      isSearchable={false}
+      unstyled
+      styles={customStyles}
+      aria-labelledby="aria-label"
+      inputId="aria-example-input"
+      className={className}
+      classNames={customClasses}
+    />
   );
 };
