+++ src/pages/NotFound.tsx
@@ -4,7 +4,7 @@ export const NotFound = () => {
     <>
       <div className="absolute bottom-0 left-0 right-0 top-0 z-0 col-span-full flex items-center justify-center">
         <h2 className="text-8xl">404 Page Not Found</h2>
-        <div className="rounded-none border border-Icons px-3 py-0 shadow-none focus-within:border-Primary hover:border-Secondary" />
+        <div className="place-self-center" />
       </div>
     </>
   );
